+++ spring-cloud-tencent-polaris-context/src/main/java/com/tencent/cloud/polaris/context/config/PolarisContextAutoConfiguration.java
@@ -48,9 +48,7 @@ public class PolarisContextAutoConfiguration {
 
 	@Bean(name = "polarisContext", initMethod = "init", destroyMethod = "destroy")
 	@ConditionalOnMissingBean
-	public SDKContext polarisContext(PolarisContextProperties properties, Environment environment,
-			List<PolarisConfigModifier> modifierList)
-			throws PolarisException {
+	public SDKContext polarisContext(PolarisContextProperties properties, Environment environment, List<PolarisConfigModifier> modifierList) throws PolarisException {
 		return SDKContext.initContextByConfig(properties.configuration(modifierList, () -> {
 			return environment.getProperty("spring.cloud.client.ip-address");
 		}));
@@ -80,7 +78,7 @@ public class PolarisContextAutoConfiguration {
 	}
 
 	@Bean
-	public ServiceRuleManager serviceRuleManager(SDKContext sdkContext) {
-		return new ServiceRuleManager(sdkContext);
+	public ServiceRuleManager serviceRuleManager(SDKContext sdkContext, ConsumerAPI consumerAPI) {
+		return new ServiceRuleManager(sdkContext, consumerAPI);
 	}
 }