+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - [refactor:update to junit 5.](https://github.com/Tencent/spring-cloud-tencent/pull/866)
 - [docs:support auto snapshot release in GitHub Action.](https://github.com/Tencent/spring-cloud-tencent/pull/871)
 - [feature:add User-Agent:polaris for healthyCheck api.](https://github.com/Tencent/spring-cloud-tencent/pull/872)
+- [optimize:optimize ServiceRuleManager.](https://github.com/Tencent/spring-cloud-tencent/pull/869)